The work I've done so far
I created a random DNA sequence generator function in order to test
future functions out. I've also made a DNA->RNA and RNA->Protein
functions. The most complex function so far is the longestORF finder.
This finds the longest open reading frame. This likely should be edited
in the future because it is pretty messy, I just added functionality to
make it work (not to make it readable). In the driver, I am currently
testing a large DNA sequence from bacteria where I can test if the ORF
finder is working. I know this because I use the NCBI website for ORF
finding --essentially a program's function that I'm imitating in this
stage of the project.